Address PR review feedback
- Rename resolveFinishRoute to getFinishRoute
- Combine when cases returning the same value
- Inline assetId.toIdentifier() in getByAssetId
- Rename dstValidator to destinationValidator
- Move GemGasPriceType extensions to ext/ layer
- Use val json pattern in assembleMetadata
